Move dialog to table content
